pyproject.toml
[project]
name = "tfg-arquitecturas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "albumentations>=2.0.8",
    "einops>=0.8.2",
    "hydra-core>=1.3.2",
    "numpy>=2.4.4",
    "omegaconf>=2.3.0",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "rich>=15.0.0",
    "timm>=1.0.26",
    "torch>=2.11.0",
    "torchmetrics>=1.9.0",
    "torchvision>=0.26.0",
]
